Modify username in calamares config on installer images
When the username is configured in the image calamares does not know
which user to change the password of, causing the install to fail.

This cannot be cleanly done live, as the username has to be discovered
from uid and calamares has to modify its own configuration file.

Solves https://gitlab.com/mobian1/calamares-settings-mobian/-/issues/2