Remove gnome-authenticator
The maintainer of gnome-authenticator has appraoched me to have gnome-authenticator removed
from the default set of applications. 3.32.2+dfsg1-2 is so old, that he does not want to see
it installed by default, and any newer version depends on gtk4 so is not available for us.

It will always be installable through Debian for those in need. My heart is bleeding removing
it by default, but I don't feel we should include applications by default against the express
desire of their maintainers.