rootfs: change the archive name when using non-free components
The rootfs archive was the same no matter whether it was configured to 
use non-free packages from the Debian archive or not. This led to issues 
when building for example a PinePhone image then a OnePlus6 image, as 
the latter would re-use the previously generated rootfs, which isn't 
configured for non-free (and thus the OP6 build would fail).

Having separate rootfs archives for "normal" and "non-free" images 
solves this issue.