Add CI for Mobian images
This commit adds a full CI pipeline for building Mobian images.

This pipeline relies on a specific runner which is a privileged docker
runner on a server we have full control over. This is required for debos
to run properly as it needs to access the kvm device.

As gitlab.com doesn't allow artifacts over 1GB we also implement a
custom artifact storage process, which can work only because we use our
own runner, as previously mentioned.

Resulting images are then uploaded to:
- https://ci-images.mobian-project.org/<DATE.TIME> for normal pipelines
- https://images.mobian-project.org/<DEVICE>/nightly for scheduled
  pipelines

Notes:
- Cleaning up old images will be executed manually for the time being,
  but automating this process should be considered for a follow-up task.
- Installer images are always uploaded to ci-images as they require more
  testing before being released.