rootfs-device: enable non-free repo if not done before
Mobian non-free repo is always enabled, however some devices require
firmware packages only available in Debian non-free. So far we enabled
it when building the base rootfs, but it requires building different
rootfs archives for devices needing non-free and other ones. The problem
is that, aside from the apt source configuration, both rootfs are
strictly identical, building 2 of those is therefore a waste of time and
resources.

By making it possible to enable non-free while building the
device-specific rootfs we can optimize part of this process.