devices: stop installing u-boot on devices supported by Tow-Boot
As we move towards generic per-architecture (instead of per-device)
images, installing u-boot on images becomes an annoyance. Let's drop
it entirely for devices supported by Tow-Boot, so we don't have to care
about it anymore.

As a consequence, it isn't necessary to reserve space for the bootloader
anymore, and we can have partitions start at the beginning of the drive
by default (only the Librem 5 still needs the bootloader in the image,
but we can expect it to change at some point).