recipes: split out base packages and Phosh-related ones
In order to make it easier to add other environments in the (hopefully
near) future, this commit ensures the base packages (not tied to any
graphical environment) are installed separately. That way we can add
`packages-<env>.yaml` includes without having to worry about low-level
dependencies.

While doing so, this commit renames `kgx` to `gnome-console`, reflecting
changes both upstream and in Debian, and drops `gstreamer1.0-gtk3`: this
package was initially installed for `fractal`, which isn't shipped by
default (nor in our repo) anymore.