image: force rebuild initramfs after deploying to image
Without a valid `/etc/fstab` and partitioning, the initramfs generated
during the previous stages has issues and generally won't boot. Make
sure it's rebuilt once we deployed the filesystem to the image.

This command was previously run in the `bootloader.sh` script for the
`sdm845` family, remove it now as it would be redundant.