Rework recipes for using a single rootfs for non-installer images
Creating a per-arch "base rootfs", then a per-device "final rootfs" and
finally the image is only useful on devices for which we also build an
installer image. Even there, it complicates `image.yaml` due the need to
get rid of FDE-related packages and ultimately rebuild the initramfs.

Using directly the base rootfs for all images reduces one step and
should help speed up image generation. The time "lost" by installing
device-specific packages twice (when generating the "normal" image and
the installer one) on some devices is low enough to be disregarded.