console: move console-setup.sh to each device dir
Duplicates `console-setup.sh` under each family device directory. The
purpose is to avoid the usage of relative paths `../../` in the recipe,
so we put the script next to each device recipe.
This unfortunately solution is imposed by the limitations of `debos`
when we run the script in a `chroot` context.