DO NOT MERGE Add generic arm64 image
Those changes allow building an image able to boot on several arm64
devices, as long as a boot firmware, such as Tow-Boot for the PinePhone
(Pro), is already installed on the device.

It relies on our generic kernel (`linux-image-5.17-arm64-mobian`) and
`mobian-tweaks-data`/`tweakster` for configuring the system at boot time
based on the exact device being used.

*** IMPORTANT ***

Please note `mobian-tweaks-data` is a temporary package for testing
purposes only! It isn't meant to be carried as part of Mobian, but
should instead replace Debian's `mobile-tweaks` package.