gitlab-ci: update for move to salsa
By moving to salsa, we need to change a few parameters (such as the
default docker image and root namespace).

We're also improving project security by using the less-privileged
`gitlab-ci` user instead of `builder` for all tasks except the weekly
images publishing. This ensures the SSH key used for publishing can be a
protected CI variable, and therefore more secure as it won't be
available from unmerged branches.