switch to systemd-repart for resizing
This requires moving to GPT based images which on u-boot based devices
requires removal of the ESP for extlinux to work. As the new repart
config is simple and only used once, it is added as an overlay rather
than as a package.

x-systemd.growfs flags are also added to the fstab entries for both the
rootfs and the installer rootfs. This allows the existing installer to
conduct the resize without modification.