librem5: bootloader: ensure `systemd-repart` won't overwrite u-boot
`systemd-repart` wants to discard all sectors between the end of the GPT
and the start of the 1st partition. On this device, this space is where
the bootloader is stored, ending up with a non-booting device once
`systemd-repart` has been executed.

In order to avoid this issue, we must therefore create a new (hidden)
protective partition to ensure `systemd-repart` won't touch this area.