librem5: bootloader: simplify image creation
The current L5 images involves a few complex steps, especially for
creating the bootloader protective partition. This originated from the
belief that `u-boot` had to be flashed at offset 2KiB, requiring a GPT
resizing to make it small enough to fit in the first 2K of the image
(among other hacks).

However this belief is actually wrong: Purism does provide an image
(including a binary aimed at the M4 core) to be flashed at this offset,
but the bootloader itself must be flashed at offset 0x8400 (33KiB), as
specified in the i.MX8MQ TRM.

We can therefore get rid of the complicated `gdisk`-based script and
get back to using the good ol' `parted` for this task.