Rename `phosh` images to `graphical`
For a while we considered creating one image per graphical environment.
This was mostly due to the lack of a proper display manager usable on
mobile, touch-only devices.

As this is no longer the case, we can now ship a single image containing
multiple graphical environments. As a consequence, all explicit
references to `phosh` (except the corresponding packages, of course) are
dropped in favor of `graphical`, which is more generic. This leaves room
for a `console` environment as well.