devices: introduce generic sdm845 images
Now that `droid-juicer` is in Debian and works as expected, we can get
rid of per-device firmware and ship a single rootfs for all SDM845-based
devices we support. The only remaining device-specific image is the
boot image, which cannot be generic as it embeds the DTB for the
device it targets.