librem5: bootloader: flash u-boot from the "distroboot" MR
Current Purism u-boot is failing when it finds an `extlinux.conf` file,
as it lacks some plumbing for dealing with compressed kernels. Let's use
the CI artifact from
https://source.puri.sm/Librem5/uboot-imx/-/merge_requests/55 instead, as
it will work better for our use-case.

That way we can also avoid calling the `zz-sync-dtb` script as
generating a valid `extlinux.conf` will be enough now.